refactor: Extract shared integration test suite into cloud_storage_integration_test_support package
The Dart package validator rejects importing `test` from `lib/` when
it is only a dev dependency, but promoting it to a regular dependency
would leak it to all consumers of `serverpod_cloud_storage_s3_compat`.

- Create a local-only (`publish_to: none`) package that owns the
  `test` dependency and hosts `runS3CompatIntegrationTests` and
  `S3CompatTestConfig`
- Move the shared test suite out of `serverpod_cloud_storage_s3_compat/lib/`
- Update consumer packages (S3, R2, GCP) to import from the new package
- Update INTEGRATION_TESTING.md to reflect the new structure
