+++ integrations/cloud_storage_integration_test_support/lib/cloud_storage_integration_test_support.dart
@@ -1,0 +1,4 @@
+library;
+
+export 'src/run_s3_compat_integration_tests.dart';
+export 'src/s3_compat_test_config.dart';
