+++ templates/pubspecs/integrations/serverpod_cloud_storage_gcp/pubspec.yaml
@@ -22,6 +22,8 @@ dependencies:
   serverpod_cloud_storage_s3_compat: SERVERPOD_VERSION
 
 dev_dependencies:
+  cloud_storage_integration_test_support:
+    path: ../cloud_storage_integration_test_support
   lints: ">=3.0.0 <7.0.0"
   mocktail: ^1.0.0
   serverpod_client: SERVERPOD_VERSION
