+++ templates/pubspecs/integrations/serverpod_cloud_storage_r2/pubspec.yaml
@@ -15,6 +15,8 @@ dependencies:
   serverpod_cloud_storage_s3_compat: SERVERPOD_VERSION
 
 dev_dependencies:
+  cloud_storage_integration_test_support:
+    path: ../cloud_storage_integration_test_support
   lints: ">=3.0.0 <7.0.0"
   test: ^1.25.5
